requirements.txt
# 核心依赖
eventlet==0.38.0
Flask==3.1.0
Flask_SocketIO==5.4.1
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7

# 数据处理与分析
matplotlib==3.9.2
neurokit2==0.2.10
numpy~=1.26.0
pandas==2.2.3
scipy==1.13.0
biosppy==1.0.0  # 生物信号处理
wfdb==4.1.2  # 波形数据库
python-dateutil==2.8.2  # 日期时间处理

# 设备通信
pyserial==3.5  # 串口通信
python-socketio==5.11.4  # Socket.IO客户端
# 蓝牙通信（可选，需要单独安装）
# 安装方法：
# 1. 安装系统依赖: sudo apt install libbluetooth-dev python3-dev
# 2. 安装PyBluez: sudo pip3 install git+https://github.com/pybluez/pybluez.git#egg=pybluez

# 数据库支持
influxdb-client==1.36.1  # 时序数据库
pymongo==4.5.0  # 文档数据库
redis==5.0.0  # 缓存

# 安全与认证
Flask-Login==0.6.3  # 用户认证
Flask-Bcrypt==1.0.1  # 密码哈希
pyjwt==2.7.0  # JWT认证
passlib==1.7.4  # 密码哈希库

# API与通信
Flask-RESTful==0.3.10  # RESTful API支持
Flask-Cors==4.0.0  # 跨域支持
requests==2.31.0  # HTTP客户端
pika==1.3.2  # RabbitMQ客户端

# 前端与UI
Flask-Assets==2.1.0  # 静态资源管理

# 测试与开发
pytest==7.4.0  # 单元测试
pytest-cov==4.1.0  # 测试覆盖率
flake8==6.1.0  # 代码风格检查

# 其他工具
tqdm==4.66.1  # 进度条
python-dotenv==1.0.0  # 环境变量管理
loguru==0.7.2  # 日志管理
fpdf==1.7.2  # PDF生成
minify==0.4.0  # 前端资源优化
